backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0

# STM32串口通信
pyserial==3.5

# WebSocket
websockets==12.0

# AI API
openai==1.6.0
httpx==0.25.2

# 环境变量
python-dotenv==1.0.0

# 工具
numpy==1.26.0